Add ranking to api
